Removed FixtureGenerationTest
